[IMP] point_of_sale: fields added to X and Z reports The module pos_daily_sales_reports has been added in 16.0 and forward-ported in master. In master we merge this module with the module point_of_sale.
Forward-port PR: odoo/odoo#113228
