+++ addons/point_of_sale/security/ir.model.access.csv
@@ -51,3 +51,4 @@ access_pos_close_session_wizard,access.pos.close.session.wizard,model_pos_close_
 access_account_cash_rounding_pos_user,account.cash.rounding pos_user,account.model_account_cash_rounding,group_pos_user,1,0,0,0
 access_pos_bill,pos.bill pos_user,model_pos_bill,group_pos_user,1,1,1,1
 access_pos_session_check_product_wizard,access.pos.session.check.product.wizard,model_pos_session_check_product_wizard,group_pos_user,1,1,1,0
+access_pos_daily_sales_reports_wizard,access.pos.daily.sales.reports.wizard,model_pos_daily_sales_reports_wizard,group_pos_manager,1,1,1,0
