+++ addons/point_of_sale/wizard/__init__.py
@@ -5,3 +5,4 @@ from . import pos_details
 from . import pos_payment
 from . import pos_close_session_wizard
 from . import pos_session_check_product_wizard
+from . import pos_daily_sales_reports
