+++ addons/point_of_sale/wizard/pos_daily_sales_reports.py
