+++ addons/point_of_sale/wizard/pos_daily_sales_reports.xml
@@ -16,7 +16,7 @@
     </record>
 
     <record id="action_report_pos_daily_sales_reports" model="ir.actions.act_window">
-        <field name="name">Daily Reports</field>
+        <field name="name">Session Report</field>
         <field name="res_model">pos.daily.sales.reports.wizard</field>
         <field name="view_mode">form</field>
         <field name="target">new</field>
