--- addons/pos_daily_sales_reports/__init__.py
@@ -1,2 +1,0 @@
-from . import models
-from . import wizard
