--- addons/pos_daily_sales_reports/wizard/__init__.py
@@ -1,1 +1,0 @@
-from . import pos_daily_sales_reports_wizard
